Added new object classes: IntersectionPoint, Room, RoomProbs, RoomProbsPerPoint. Added class: IntersectsToRooms;
IntersectionPoint - this class is an object wrapper for each intersection that has x,y, and confidence parameters.
Room - Made this class public and added support for both the old and new style lookup table. 
RoomProbs - this object wrapper holds the room id, distance values, and confidence that this is the true room of the parent point.
RoomProbsPerPoint - object wrapper holds the IntersectionPoint object and an ArrayList of the probabilities for each room being the true room

IntersectsToRooms - this class actually calculates the distance from each intersection to each room and uses the distance and confidence to create an overall confidence for each intersection